% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_H2test}
\alias{data_H2test}
\title{Data very simple structure}
\format{An object of class \code{list} of length 4.}
\usage{
data_H2test
}
\description{
Data randomly generated for simple H2 test
}
\keyword{datasets}
