% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fox_Raoul2010}
\alias{fox_Raoul2010}
\title{Data of red fox (Vulpes vulpes) predation on 2 vole prey species (Arvicola scherman and Microtus arvalis)}
\format{An object of class \code{list} of length 5.}
\usage{
fox_Raoul2010
}
\description{
Data randomly generated
}
\keyword{datasets}
