% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophicFit.R
\name{trophicFit}
\alias{trophicFit}
\title{Fitting models for functional responses using Bayesian inference}
\usage{
trophicFit(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{\dots}{Further arguments to be passed to generic methods}
}
\description{
Fitting functional reponses using Bayesian inference for objects.
}
