% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophicFit.R
\name{trophicFit_H2test}
\alias{trophicFit_H2test}
\title{Function to fit H2test model}
\usage{
trophicFit_H2test(data, ...)
}
\arguments{
\item{data}{an object used to select a method}
}
\description{
Estimate Bayesian Holling type 2 model using Stan on a single species
dataset
}
