\name{trophicR-package}
\alias{trophicR-package}
\alias{trophicR}
\docType{package}
\title{
\packageTitle{trophicR}
}
\description{
\packageDescription{trophicR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{trophicR}
\packageIndices{trophicR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{trophicR}

Maintainer: \packageMaintainer{trophicR}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package 
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
