% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophicFit.R
\name{trophicFit}
\alias{trophicFit}
\title{Function to fit Holling models}
\usage{
trophicFit(data, trophic_model = NULL, ...)
}
\arguments{
\item{data}{an object used to select a method}
}
\description{
Estimate Bayesian Holling type 2 or 3 model using Stan
}
